requirements.txt
azure-functions
azure-storage-blob
openai
requests
pillow  # For image processing in case it's used for the input images
